refactor(speaker): clean up the bgp part
