+++ pkg/speaker/subnet.go
@@ -4,17 +4,13 @@ package speaker
 import (
 	"context"
 	"fmt"
-	"net"
-	"strconv"
 	"strings"
 
 	bgpapi "github.com/osrg/gobgp/v3/api"
 	bgpapiutil "github.com/osrg/gobgp/v3/pkg/apiutil"
-	"github.com/osrg/gobgp/v3/pkg/packet/bgp"
+
 	"github.com/vishvananda/netlink"
 	"golang.org/x/sys/unix"
-	"google.golang.org/protobuf/types/known/anypb"
-	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/klog/v2"
 
@@ -23,31 +19,16 @@ import (
 )
 
 const (
-	// "cluster" is the default policy
+	// announcePolicyCluster makes the Pod IPs/Subnet CIDRs be announced from every speaker, whether there's Pods
+	// that have that specific IP or that are part of the Subnet CIDR on that node. In other words, traffic may enter from
+	// any node hosting a speaker, and then be internally routed in the cluster to the actual Pod. In this configuration
+	// extra hops might be used. This is the default policy to Pods and Subnets.
 	announcePolicyCluster = "cluster"
-	announcePolicyLocal   = "local"
+	// announcePolicyLocal makes the Pod IPs be announced only from speakers on nodes that are actively hosting
+	// them. In other words, traffic will only enter from nodes hosting Pods marked as needing BGP advertisement,
+	// or Pods with an IP belonging to a Subnet marked as needing BGP advertisement. This makes the network path shorter.
+	announcePolicyLocal = "local"
 )
-
-func isPodAlive(p *v1.Pod) bool {
-	if p.Status.Phase == v1.PodSucceeded && p.Spec.RestartPolicy != v1.RestartPolicyAlways {
-		return false
-	}
-
-	if p.Status.Phase == v1.PodFailed && p.Spec.RestartPolicy == v1.RestartPolicyNever {
-		return false
-	}
-
-	if p.Status.Phase == v1.PodFailed && p.Status.Reason == "Evicted" {
-		return false
-	}
-	return true
-}
-
-func isClusterIPService(svc *v1.Service) bool {
-	return svc.Spec.Type == v1.ServiceTypeClusterIP &&
-		svc.Spec.ClusterIP != v1.ClusterIPNone &&
-		len(svc.Spec.ClusterIP) != 0
-}
 
 func (c *Controller) syncSubnetRoutes() {
 	maskMap := map[string]int{kubeovnv1.ProtocolIPv4: 32, kubeovnv1.ProtocolIPv6: 128}
@@ -214,150 +195,4 @@ func (c *Controller) syncSubnetRoutes() {
 			}
 		}
 	}
-}
-
-func routeDiff(expected, exists []string) (toAdd, toDel []string) {
-	expectedMap, existsMap := map[string]bool{}, map[string]bool{}
-	for _, e := range expected {
-		expectedMap[e] = true
-	}
-	for _, e := range exists {
-		existsMap[e] = true
-	}
-
-	for e := range expectedMap {
-		if !existsMap[e] {
-			toAdd = append(toAdd, e)
-		}
-	}
-
-	for e := range existsMap {
-		if !expectedMap[e] {
-			toDel = append(toDel, e)
-		}
-	}
-	return toAdd, toDel
-}
-
-func parseRoute(route string) (string, uint32, error) {
-	var prefixLen uint32 = 32
-	prefix := route
-	if strings.Contains(route, "/") {
-		prefix = strings.Split(route, "/")[0]
-		strLen := strings.Split(route, "/")[1]
-		intLen, err := strconv.Atoi(strLen)
-		if err != nil {
-			return "", 0, err
-		}
-		prefixLen = uint32(intLen)
-	}
-	return prefix, prefixLen, nil
-}
-
-func (c *Controller) addRoute(route string) error {
-	routeAfi := bgpapi.Family_AFI_IP
-	if util.CheckProtocol(route) == kubeovnv1.ProtocolIPv6 {
-		routeAfi = bgpapi.Family_AFI_IP6
-	}
-
-	nlri, attrs, err := c.getNlriAndAttrs(route)
-	if err != nil {
-		return err
-	}
-	for _, attr := range attrs {
-		_, err = c.config.BgpServer.AddPath(context.Background(), &bgpapi.AddPathRequest{
-			Path: &bgpapi.Path{
-				Family: &bgpapi.Family{Afi: routeAfi, Safi: bgpapi.Family_SAFI_UNICAST},
-				Nlri:   nlri,
-				Pattrs: attr,
-			},
-		})
-		if err != nil {
-			klog.Errorf("add path failed, %v", err)
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *Controller) getNlriAndAttrs(route string) (*anypb.Any, [][]*anypb.Any, error) {
-	neighborAddresses := c.config.NeighborAddresses
-	if util.CheckProtocol(route) == kubeovnv1.ProtocolIPv6 {
-		neighborAddresses = c.config.NeighborIPv6Addresses
-	}
-
-	prefix, prefixLen, err := parseRoute(route)
-	if err != nil {
-		return nil, nil, err
-	}
-	nlri, _ := anypb.New(&bgpapi.IPAddressPrefix{
-		Prefix:    prefix,
-		PrefixLen: prefixLen,
-	})
-
-	attrs := make([][]*anypb.Any, 0, len(neighborAddresses))
-	for _, addr := range neighborAddresses {
-		a1, _ := anypb.New(&bgpapi.OriginAttribute{
-			Origin: 0,
-		})
-		a2, _ := anypb.New(&bgpapi.NextHopAttribute{
-			NextHop: c.getNextHopAttribute(addr),
-		})
-		attrs = append(attrs, []*anypb.Any{a1, a2})
-	}
-
-	return nlri, attrs, err
-}
-
-func (c *Controller) delRoute(route string) error {
-	routeAfi := bgpapi.Family_AFI_IP
-	if util.CheckProtocol(route) == kubeovnv1.ProtocolIPv6 {
-		routeAfi = bgpapi.Family_AFI_IP6
-	}
-
-	nlri, attrs, err := c.getNlriAndAttrs(route)
-	if err != nil {
-		return err
-	}
-	for _, attr := range attrs {
-		err = c.config.BgpServer.DeletePath(context.Background(), &bgpapi.DeletePathRequest{
-			Path: &bgpapi.Path{
-				Family: &bgpapi.Family{Afi: routeAfi, Safi: bgpapi.Family_SAFI_UNICAST},
-				Nlri:   nlri,
-				Pattrs: attr,
-			},
-		})
-		if err != nil {
-			klog.Errorf("del path failed, %v", err)
-			return err
-		}
-	}
-	return nil
-}
-
-func getNextHopFromPathAttributes(attrs []bgp.PathAttributeInterface) net.IP {
-	for _, attr := range attrs {
-		switch a := attr.(type) {
-		case *bgp.PathAttributeNextHop:
-			return a.Value
-		case *bgp.PathAttributeMpReachNLRI:
-			return a.Nexthop
-		}
-	}
-	return nil
-}
-
-func (c *Controller) getNextHopAttribute(neighborAddress string) string {
-	nextHop := c.config.RouterID
-	routes, err := netlink.RouteGet(net.ParseIP(neighborAddress))
-	if err == nil && len(routes) == 1 && routes[0].Src != nil {
-		nextHop = routes[0].Src.String()
-	}
-	proto := util.CheckProtocol(nextHop)
-	nextHopIP := net.ParseIP(nextHop)
-	nodeIP := c.config.NodeIPs[proto]
-	if nodeIP != nil && nextHopIP != nil && nextHopIP.Equal(c.config.PodIPs[proto]) {
-		nextHop = nodeIP.String()
-	}
-	return nextHop
 }
